requirements.txt
setuptools~=45.2.0
numpy~=1.18.3
pygmsh~=6.1.1
scipy~=1.4.1
PyYAML ~= 5.3.1
meshmagick~=2.0